requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
plotly==5.18.0
openai==1.13.3
requests==2.31.0
python-binance==1.0.19
yfinance==0.2.36
tweepy==4.14.0
ccxt==4.2.24
python-dotenv==1.0.1
SQLAlchemy==2.0.27
pydantic==2.6.1
bcrypt==4.1.2
cryptography==42.0.2
fastapi==0.109.2
uvicorn==0.27.1
PyJWT==2.10.1
pytest==7.4.3
pytest-mock==3.12.0